requirements.txt
# Base Libraries
python-dotenv==1.0.0
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
Werkzeug==2.3.7
blinker==1.6.2
sqlalchemy==2.0.21
WTForms==3.0.1
dnspython==2.4.2
idna==3.4
packaging==23.2
alembic==1.12.0
setproctitle==1.3.3
paho-mqtt==1.6.1

# Main Libraries
Flask==2.3.3
PyMySQL==1.1.0
email-validator==2.0.0.post2
mariadb==1.1.7
gunicorn==21.2.0

# Extensions and Libraries that depend on the main libraries
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6